MULTI-Z
plane shizz

hmm it works

multiz cables and pipes

monsterstation

oops

changes shizz

more shizz

adds like a nanite lab or something

fuck gotta work on other shit

more shizz

minor fixes

some work on engi

improvements

more map stuff

More monsterstation changes + multiz UI icons

A whole bunch of shit

hm

help I not good at code

whole bunch of fixes

+++ code/controllers/subsystem/throwing.dm
@@ -146,6 +146,10 @@ SUBSYSTEM_DEF(throwing)
 
 	if (callback)
 		callback.Invoke()
+	
+	var/turf/T = get_turf(thrownthing)
+	if(istype(T))
+		T.zFall(thrownthing) // zFall is disabled during throws so that things can be thrown over gaps, but we still want to be able to throw stuff down.
 
 	qdel(src)
